+++ src/main/java/mb/minecraft/dao/VillagerDao.java
@@ -11,7 +11,7 @@ import mb.minecraft.model.Villager;
  * @author mikebro
  */
 public interface VillagerDao extends DisposableBean {
-	public Villager selectOneById( Long id );
+	public Villager selectOneById( int id );
 	public Villager selectOneByName( String name );
 	public List<Villager> selectAll();
 	public Villager insertOne( Villager villager );
